+++ source/isaaclab_tasks/isaaclab_tasks/manager_based/manipulation/assemble_trocar/__init__.py
@@ -42,3 +42,12 @@ gym.register(
     },
     disable_env_checker=True,
 )
+
+gym.register(
+    id="Isaac-Assemble-Trocar-G129-Dex3-RLinf-MultiModal-v0",
+    entry_point="isaaclab.envs:ManagerBasedRLEnv",
+    kwargs={
+        "env_cfg_entry_point": g129_dex3_env_cfg.G1AssembleTrocarMultiModalEnvCfg,
+    },
+    disable_env_checker=True,
+)